# api/openapi.yaml
openapi: 3.0.3
info:
  title: Self-Hosted RWA Platform API
  version: 1.0.0
  description: >
    FROZEN API skeleton (lead-owned). State-changing endpoints accept an
    Idempotency-Key header. Errors use stable machine-readable codes.
    On-chain actions (investor buy/request/claim/cancel, admin/treasury/redemption
    wallet actions) are encoded client-side by the SPA and broadcast from the
    connected wallet — never executed by a custodial server key.
servers:
  - url: http://localhost:8080
security:
  - BearerJwt: []
...
paths:
...
  /api/v1/compliance/wallets:
    get:
      tags: [compliance]
      operationId: listWallets
      x-required-role: operator
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: "Wallets (bounded/paginated, see Limit/Cursor)"
          headers:
            X-Total-Count: { $ref: "#/components/headers/XTotalCount" }
            X-Page-Size: { $ref: "#/components/headers/XPageSize" }
            X-Next-Cursor: { $ref: "#/components/headers/XNextCursor" }
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/WalletStatus" } } } }
  /api/v1/compliance/challenge:
    post:
      tags: [compliance]
      operationId: createChallenge
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [address], properties: { address: { type: string } } } } }
      responses:
        "200": { description: Challenge, content: { application/json: { schema: { $ref: "#/components/schemas/Challenge" } } } }
  /api/v1/compliance/challenge/verify:
    post:
      tags: [compliance]
      operationId: verifyChallenge
      description: >
        Submit the wallet signature over a previously issued challenge to prove ownership.
        On success also mints a short-lived subject-scoped session (bound to `address`
        only) so an investor can read their own status without an operator API key.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ChallengeVerify" } } }
      responses:
        "200": { description: Verified, content: { application/json: { schema: { $ref: "#/components/schemas/VerifyChallengeResult" } } } }
        "400": { $ref: "#/components/responses/Error" }
...
  /api/v1/compliance/webhooks:
    get:
      tags: [compliance]
      operationId: listWebhookEvents
      x-required-role: operator
      description: KYC webhook history for the compliance screen.
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: "Webhook events (bounded/paginated, see Limit/Cursor)"
          headers:
            X-Total-Count: { $ref: "#/components/headers/XTotalCount" }
            X-Page-Size: { $ref: "#/components/headers/XPageSize" }
            X-Next-Cursor: { $ref: "#/components/headers/XNextCursor" }
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/WebhookEvent" } } } }
  /api/v1/audit-logs:
    get:
      tags: [compliance]
      operationId: listAuditLogs
      x-required-role: operator
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: >
            Audit log entries (bounded/paginated; the
            underlying store is a most-recent-N query so X-Total-Count is
            omitted — only X-Page-Size/X-Next-Cursor are sent).
          headers:
            X-Page-Size: { $ref: "#/components/headers/XPageSize" }
            X-Next-Cursor: { $ref: "#/components/headers/XNextCursor" }
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/AuditLog" } } } }
  /api/v1/compliance/status:
    post:
      tags: [compliance]
      operationId: setComplianceStatus
      parameters: [ { $ref: "#/components/parameters/IdempotencyKey" } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SetStatusRequest" } } }
      responses:
        "202": { description: Submitted, content: { application/json: { schema: { $ref: "#/components/schemas/TxRef" } } } }
  /api/v1/compliance/webhook:
    post:
      tags: [compliance]
      operationId: kycWebhook
      description: >
        HMAC-signed KYC provider webhook. Signature header `X-Webhook-Signature` =
        lowercase hex of HMAC-SHA256(rawBody, sharedSecret), constant-time compared.
        Replay/freshness-protected: the server requires provider + eventId
        + occurredAt, enforces uniqueness on (provider, eventId) in addition to raw-body-hash
        dedup, rejects occurredAt outside a 24h-past / 5min-future window, rejects a
        strictly-older decision for an address than the latest already-applied one, normalizes
        the address to EIP-55, and rejects negative / >100y-future validUntil before the
        int64→uint64 cast. The occurrence timestamp travels in-body, so it is covered by the
        existing HMAC-over-raw-body signature (no separate header). Not protected by ApiKeyAuth.
      security: []
      parameters:
        - { name: X-Webhook-Signature, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/WebhookPayload" } } }
      responses:
        "202": { description: "Accepted and durably queued. The decision is stored and applied on-chain asynchronously by the reconciler; it is NOT yet applied when this returns. Poll the webhook events list for applyStatus." }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "409": { description: "Replay/stale/out-of-order decision (same code family as replay).", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /api/v1/sales/purchases:
    get:
      tags: [sales]
      operationId: listPurchases
      x-required-role: operator
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: "Purchases (bounded/paginated, see Limit/Cursor)"
          headers:
            X-Total-Count: { $ref: "#/components/headers/XTotalCount" }
            X-Page-Size: { $ref: "#/components/headers/XPageSize" }
            X-Next-Cursor: { $ref: "#/components/headers/XNextCursor" }
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Purchase" } } } }
...
components:
  securitySchemes:
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Admin JWT (HMAC-signed) issued by POST /auth/session after a successful
        wallet-signature challenge whose signer equals the configured project admin
        address. Attach as `Authorization: Bearer <jwt>`. Stateless — no server-side
        revocation; logout drops the token client-side. Investor read routes use
        X-Wallet-Session instead; requests with no/invalid token fall through to the
        public (read-only) surface.
    WalletSession:
      type: apiKey
      in: header
      name: X-Wallet-Session
      description: >
        Short-lived (≈15 min) opaque session token minted at challenge-verify, scoped to
        a single wallet address. Grants read access to that address's own status only —
        never an admin capability. Kept separate from the admin BearerJwt so the two
        auth models never collide.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: >
        Required on every side-effecting endpoint — a missing or optional key
        would let a mutation bypass idempotency protection entirely. Scoped
        server-side to the authenticated caller's
        credential identity, the operation, and its full concrete target
        (path parameters + normalized query), not merely the route
        template — see api.go's Idempotency middleware.
      required: true
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: >
        Max items to return. Server-enforced hard maximum
        applies even if omitted or set higher than the maximum.
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: >
        Opaque pagination cursor from a previous response's X-Next-Cursor
        header. Omit to fetch the first page.
      required: false
      schema: { type: string }
  headers:
    XTotalCount:
      description: Total items available, when cheaply known.
      schema: { type: integer }
    XPageSize:
      description: Items actually returned in this response.
      schema: { type: integer }
    XNextCursor:
      description: >
        Pass as `cursor` to fetch the next page; absent once this was the
        last page.
      schema: { type: string }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object }
...
    WalletStatus:
      type: object
      properties:
        address: { type: string }
        status: { type: string, enum: [Unknown, Allowed, Blocked] }
        validUntil: { type: integer }
        ownershipVerified: { type: boolean }
    VerifyChallengeResult:
      type: object
      description: WalletStatus for the verified address plus a short-lived subject-scoped session.
      properties:
        address: { type: string }
        status: { type: string, enum: [Unknown, Allowed, Blocked] }
        validUntil: { type: integer }
        ownershipVerified: { type: boolean }
        sessionToken: { type: string, description: "Opaque bearer for X-Wallet-Session; scoped to this address only." }
        sessionExpiresAt: { type: string, format: date-time, description: "Session expiry (≈15 min TTL)." }
...
    Challenge:
      type: object
      properties:
        address: { type: string }
        nonce: { type: string }
        message: { type: string }
        expiresAt: { type: string }
    SetStatusRequest:
      type: object
      required: [address, status]
      properties:
        address: { type: string }
        status: { type: string, enum: [Unknown, Allowed, Blocked] }
        validUntil: { type: integer }
...
    ChallengeVerify:
      type: object
      required: [address, nonce, signature]
      properties:
        address: { type: string }
        nonce: { type: string }
        signature: { type: string }
    WebhookPayload:
      type: object
      required: [provider, eventId, occurredAt, address, status]
      properties:
        provider: { type: string, description: "KYC provider id; part of the (provider,eventId) uniqueness key." }
        eventId: { type: string, description: "Provider-assigned unique event id; part of the uniqueness key." }
        occurredAt: { type: integer, format: int64, description: "Provider decision time, unix seconds. Rejected if outside 24h-past / 5min-future, or older than the latest applied decision for this address." }
        address: { type: string, description: "Subject wallet; normalized to EIP-55 checksum server-side." }
        status: { type: string, enum: [Allowed, Blocked, Pending], description: "Compliance decision." }
        validUntil: { type: integer, format: int64, description: "Expiry (unix seconds) for an Allowed decision; 0 = no expiry. Negative or >100y-future is rejected before the int64→uint64 cast." }
    WebhookEvent:
      type: object
      properties:
        id: { type: string }
        eventId: { type: string }
        address: { type: string }
        provider: { type: string }
        outcome: { type: string, enum: [Allowed, Blocked, Pending] }
        occurredAt: { type: string, format: date-time }
        receivedAt: { type: string }
        applyStatus: { type: string, enum: [Accepted, Applying, Applied, Superseded, Failed, Recorded], description: "Durable inbox/outbox state. Applied only after the on-chain setStatus tx confirms." }
        applied: { type: boolean, description: "Derived convenience flag: applyStatus == Applied." }
    AuditLog:
      type: object
      properties:
        id: { type: string }
        category: { type: string }
        actor: { type: string }
        action: { type: string }
        target: { type: string }
        createdAt: { type: string }
        details: { type: object }
    Purchase:
      type: object
      properties:
        txHash: { type: string }
        buyer: { type: string }
        recipient: { type: string }
        tokenAmount: { type: string }
        quoteAmount: { type: string }
        blockNumber: { type: integer }
        confirmations: { type: integer }
    TxRef:
      type: object
      properties:
        txHash: { type: string }
        status: { type: string }
        idempotencyKey: { type: string }
...